Enable PHPStan in tests

+++ tests/Doctrine/Tests/DBAL/Tools/DumperTest.php
@@ -100,8 +100,8 @@ class DumperTest extends DbalTestCase
         $print_r_class    = print_r($class, true);
         $print_r_expected = print_r($expected, true);
 
-        $print_r_class    = substr($print_r_class, strpos($print_r_class, '('));
-        $print_r_expected = substr($print_r_expected, strpos($print_r_expected, '('));
+        $print_r_class    = substr($print_r_class, (int) strpos($print_r_class, '('));
+        $print_r_expected = substr($print_r_expected, (int) strpos($print_r_expected, '('));
 
         self::assertSame($print_r_class, $print_r_expected);
